Rename Containers and added files back
Renamed the Containers folder to the Components folder because it is presentation

Added back containers folder and added 3 files: AddedMessage.js, MessagesList.js, and Sidebar.js file

Added code for the above files  to connect those corresponding files in the component folder to the app.
