+++ src/components/MessagesList.js
@@ -1,0 +1,27 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+const MessagesList = ({ messages }) => (
+  <section id="messages-list">
+    <ul>
+      {messages.map(message => (
+        <messages
+          key={message.id}
+          {...message}
+          />
+      ))}
+    </ul>
+  </section>
+)
+
+MessagesList.PropTypes = {
+  messages: PropTypes.arrayOf(
+    PropTypes.shape({
+      id: PropTypes.number.isRequired,
+      message: PropTypes.string.isRequired,
+      author: PropTypes.string.isRequired
+    }).isRequired
+  ).isRequired
+}
+
+export default MessagesList
