+++ src/components/Sidebar.js
@@ -4,12 +4,12 @@ import PropTypes from 'prop-types';
 const Sidebar = ({ users }) => (
   <aside id="sidebar" className="sidebar">
     <ul>
-      {users.map(user => {
+      {users.map(user=>{
         <li key={user.id}>{user.name}</li>
       })}
     </ul>
   </aside>
-)
+ )
 
 Sidebar.PropTypes = {
   users: PropTypes.arrayOf(
