+++ src/containers/AddMessage.js
@@ -1,28 +1,11 @@
-import React from 'react';
-import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import AddMessageComponent from '../components/AddMessage';
+import { addMessage } from '../actions';
 
-const AddMessage = (props) => {
-  let input
+const mapDispatchToProps = dispatch => ({
+  dispatch: (message, author) => {
+    dispatch(addMessage(message, author))
+  }
+})
 
-  return (
-    <section id="new-message"
-      <input
-      onKeyPress={(e) => {
-        if (e.key === 'Enter') {
-          props.dispatch(input.value, 'Me')
-          input.value = ''
-        }
-      }}
-      type='text'
-      ref{(node) =>
-        input = node
-      }}
-    </section>
-    );
-}
-
-AddMessage.propTypes = {
-  dispatch: PropTypes.func.isRequired
-}
-
-export default AddMessage
+export const AddMessage = connect(() => ({}), mapDispatchToProps)(AddMessageComponent)
